HNepal App/corona/requirements.txt
django
django-rest-framework

